Refactor sai_api_gen.py.
