+++ dash-pipeline/SAI/utils/dash_sai/__init__.py
@@ -1,0 +1,29 @@
+#
+# Parsed SAI objects and parsers:
+#
+# The SAI objects are parsed from the P4Runtime JSON file, generated by p4 compiler, which contains the information
+# of all tables and entry information.
+# 
+# The classes below are used to parse the P4Runtime JSON file to get the key information, so we can generate the SAI
+# API headers and implementation files afterwards.
+#
+# At high level, the hiredarchy of the SAI objects is as follows:
+#
+# DASHSAIExtensions                     : All DASH SAI extensions.
+# |- SAIEnum                            : A single enum type.
+# |  |- SAIEnumMember                   : A single enum member within the enum.
+# |- SAIAPISet                          : All information for a single SAI API set, such as routing or CA-PA mapping.
+#    |- SAIAPITable                 : All information for a single SAI API table used in the API set.
+#       |- SAIAPITableKey               : Information of a single P4 table key defined in the table.
+#       |- SAIAPITableAction  <-------| : Information of a single P4 table action defined used by the table.
+#          |- SAIAPITableActionParam -| : Information of a single P4 table action parameter used by the action.
+#
+from .dash_sai_extensions import DASHSAIExtensions
+from .dash_sai_api_set import DASHAPISet
+from .sai_api_table import SAIAPITable
+from .sai_api_table_action import SAIAPITableAction
+from .sai_api_table_key import SAIAPITableKey
+from .sai_api_table_action_param import SAIAPITableActionParam
+from .sai_api_counter import SAIAPICounter
+from .sai_enum import SAIEnum
+from .sai_enum_member import SAIEnumMember
