+++ dash-pipeline/SAI/utils/p4ir/__init__.py
@@ -1,0 +1,7 @@
+#
+# P4 IR parser and analyzer:
+#
+from .p4ir_var_ref_graph import *
+from .p4ir_var_info import *
+from .p4ir_var_ref_info import *
+from .p4ir_tree import *
